Chess clock added

+++ core/src/com/emmanuelmess/simplechess/game/GameType.kt
@@ -4,6 +4,6 @@ enum class GameCategory {
     BULLET, BLITZ, RAPID, CLASSIC, CUSTOM
 }
 
-data class GameType(val name: String?, val time: Int, val timeAdded: Int, val category: GameCategory) {
+data class GameType(val name: String?, val time: Int, val timeAdded: Long, val category: GameCategory) {
     fun getString() = name ?: "$time+$timeAdded"
 }
